fix: consider whole protein when generating LBA examples
Previously, I was using only the "pocket" atoms, which only extend ≈10Å
from the ligand.  My images are bigger than this, so this was definitely
a bug.

I also updated the code to use macromol_voxelize and macromol_gym
instead of atompaint.  I think I could simplify the dependencies some
more, but this should be ok for now.